Changed class structure so that each menu get it's own class with the JFrame extended class ProgramMenu as a parameter.
Worked on the ViewUserAccountsView, now displays SuperUser and
BasicUsers. Still a work in progress.

Created the Log In Menu

Created the BasicUserMenuView, which contains 5 buttons corresponding to
it's 5 available actions

Log Out in BasicUserMenuView was implemented

Updated error labels in both Create Account and Login to appear and
disappear depending on the users input

The QuitView class now displays an updated GoodByeMessage as well as a
digital timer that will dispose the frame after counting down from 10.
(Bug: If you try to click the back button, it will still dispose the
frame after 10 seconds)

Implemented a currentUser variable along with it's getter and setter
methods in the ProgramMenu Class. This variable is updated when logging
into and out of an account.